requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-cors-headers==3.5.0
django-oauth-toolkit==1.3.2
django-rest-framework==0.1.0
django-tinymce==3.0.2
djangorestframework==3.11.1
drf-nested-routers==0.91
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
uWSGI==2.0.19.1
